moved argc/TranslateSourceList to util/SourceList
